+++ src/routes/webhooks/paddle/plus/slack.ts
@@ -1,0 +1,169 @@
+import type { TransactionCompletedEvent } from '@paddle/paddle-node-sdk';
+import { getUserId, type PaddleCustomData } from '../../../../common/paddle';
+import createOrGetConnection from '../../../../db';
+import { SubscriptionProvider, User } from '../../../../entity/user/User';
+import { logger } from '../../../../logger';
+import { concatTextToNewline } from '../../../../common/utils';
+import { webhooks } from '../../../../common/slack';
+
+export const notifyNewPaddlePlusTransaction = async ({
+  event,
+}: {
+  event: TransactionCompletedEvent;
+}) => {
+  const { data } = event;
+  const { customData, subscriptionId } = data ?? {};
+  const { user_id, gifter_id } = (customData ?? {}) as PaddleCustomData;
+  const purchasedById = gifter_id ?? user_id;
+  const subscriptionForId = await getUserId({
+    userId: user_id,
+    subscriptionId,
+  });
+  const con = await createOrGetConnection();
+  const flags = (
+    await con.getRepository(User).findOne({
+      select: ['subscriptionFlags'],
+      where: { id: subscriptionForId },
+    })
+  )?.subscriptionFlags;
+
+  if (gifter_id && !flags?.giftExpirationDate) {
+    logger.error(
+      { provider: SubscriptionProvider.Paddle, data: event },
+      'Gifted subscription without expiration date',
+    );
+  }
+
+  const origin = data?.origin;
+  const productId = data?.items?.[0].price?.productId;
+
+  const total = data?.items?.[0]?.price?.unitPrice?.amount || '0';
+  const currencyCode =
+    data?.items?.[0]?.price?.unitPrice?.currencyCode || 'USD';
+
+  const localTotal = data?.details?.totals?.total || '0';
+  const localCurrencyCode = data?.currencyCode || 'USD';
+
+  if (origin === 'subscription_recurring') {
+    return;
+  }
+
+  const headerText = (() => {
+    if (gifter_id) {
+      return 'Gift subscription :gift: :paddle:';
+    }
+
+    return 'New Plus subscriber :moneybag: :paddle:';
+  })();
+
+  const blocks = [
+    {
+      type: 'header',
+      text: {
+        type: 'plain_text',
+        text: headerText,
+        emoji: true,
+      },
+    },
+    {
+      type: 'section',
+      fields: [
+        {
+          type: 'mrkdwn',
+          text: concatTextToNewline(
+            '*Transaction ID:*',
+            `<https://vendors.paddle.com/transactions-v2/${data.id}|${data.id}>`,
+          ),
+        },
+        {
+          type: 'mrkdwn',
+          text: concatTextToNewline(
+            '*Customer ID:*',
+            `<https://vendors.paddle.com/customers-v2/${data.customerId}|${data.customerId}>`,
+          ),
+        },
+      ],
+    },
+    {
+      type: 'section',
+      fields: [
+        {
+          type: 'mrkdwn',
+          text: concatTextToNewline(
+            '*Type:*',
+            `<https://vendors.paddle.com/products-v2/${productId}|${flags?.cycle}>`,
+          ),
+        },
+        {
+          type: 'mrkdwn',
+          text: concatTextToNewline(
+            '*Purchased by:*',
+            `<https://app.daily.dev/${purchasedById}|${purchasedById}>`,
+          ),
+        },
+      ],
+    },
+    {
+      type: 'section',
+      fields: [
+        {
+          type: 'mrkdwn',
+          text: concatTextToNewline(
+            '*Cost:*',
+            new Intl.NumberFormat('en-US', {
+              style: 'currency',
+              currency: currencyCode,
+            }).format((parseFloat(total) || 0) / 100),
+          ),
+        },
+        {
+          type: 'mrkdwn',
+          text: concatTextToNewline('*Currency:*', currencyCode),
+        },
+      ],
+    },
+    {
+      type: 'section',
+      fields: [
+        {
+          type: 'mrkdwn',
+          text: concatTextToNewline(
+            '*Cost (local):*',
+            new Intl.NumberFormat('en-US', {
+              style: 'currency',
+              currency: localCurrencyCode,
+            }).format((parseFloat(localTotal) || 0) / 100),
+          ),
+        },
+        {
+          type: 'mrkdwn',
+          text: concatTextToNewline('*Currency (local):*', localCurrencyCode),
+        },
+      ],
+    },
+  ];
+
+  if (gifter_id && flags?.giftExpirationDate) {
+    blocks.push({
+      type: 'section',
+      fields: [
+        {
+          type: 'mrkdwn',
+          text: concatTextToNewline(
+            '*Gifted to:*',
+            `<https://app.daily.dev/${subscriptionForId}|${subscriptionForId}>`,
+          ),
+        },
+        {
+          type: 'mrkdwn',
+          text: concatTextToNewline(
+            '*Gift expires:*',
+            new Date(flags.giftExpirationDate).toLocaleDateString(),
+          ),
+        },
+      ],
+    });
+  }
+
+  await webhooks.transactions.send({ blocks });
+};
